create MicButton component
