+++ apps/client/src/play/Stats.tsx
@@ -59,7 +59,7 @@ function StatsMenu() {
   const averageFPS = fpsSamplesBuffer.reduce((a, b) => a + b, 0) / fpsSamplesBuffer.length;
 
   return (
-    <div className="space-y-2 rounded-lg bg-white/80 px-4 py-2 backdrop-blur">
+    <div className="space-y-2 rounded-lg bg-white/95 px-4 py-2">
       <StatDisplay label="FPS" data={Math.round(averageFPS)} />
 
       <div>
