+++ apps/client/src/play/ui/Overlay.tsx
@@ -1,6 +1,6 @@
-import { useEngine, useMic, useWebSocket } from "@wired-labs/react-client";
+import { MicButton, useEngine, useMic, useWebSocket } from "@wired-labs/react-client";
 import Link from "next/link";
-import { useRef, useState } from "react";
+import { useState } from "react";
 import { IoMdArrowRoundBack, IoMdSettings } from "react-icons/io";
 import { MdMic, MdMicOff } from "react-icons/md";
 
@@ -22,14 +22,13 @@ interface Props {
 }
 
 export default function Overlay({ id, action }: Props) {
-  const hasProducedAudio = useRef(false);
   const [openSettings, setOpenSettings] = useState(false);
 
   const engine = useEngine();
   const isMobile = useIsMobile();
   const setAvatar = useSetAvatar();
   const { send } = useWebSocket();
-  const { micEnabled, setMicEnabled, setMicTrack } = useMic();
+  const { micEnabled } = useMic();
 
   async function handleClose() {
     setOpenSettings(false);
@@ -65,24 +64,6 @@ export default function Overlay({ id, action }: Props) {
     }
   }
 
-  async function toggleMic() {
-    if (!engine) return;
-
-    // If first time using mic, request permission
-    if (!micEnabled && !hasProducedAudio.current) {
-      const stream = await navigator.mediaDevices.getUserMedia({ audio: true });
-
-      const track = stream.getAudioTracks()[0];
-      if (!track) throw new Error("No audio track found");
-
-      setMicTrack(track);
-
-      hasProducedAudio.current = true;
-    }
-
-    setMicEnabled(!micEnabled);
-  }
-
   return (
     <>
       <DialogRoot
@@ -112,12 +93,10 @@ export default function Overlay({ id, action }: Props) {
       <CrosshairTooltip action={action} />
 
       <div className="fixed top-0 right-0 z-20 space-x-2 p-4">
-        <button
-          onClick={toggleMic}
-          className="rounded-full bg-white/80 p-3 text-2xl text-neutral-900 shadow backdrop-blur-xl transition hover:bg-white/90 hover:shadow-md active:scale-95"
-        >
+        <MicButton className="rounded-full bg-white/80 p-3 text-2xl text-neutral-900 shadow backdrop-blur-xl transition hover:bg-white/90 hover:shadow-md active:scale-95">
           {micEnabled ? <MdMic /> : <MdMicOff className="text-red-700" />}
-        </button>
+        </MicButton>
+
         <button
           onClick={() => setOpenSettings(true)}
           className="rounded-full bg-white/80 p-3 text-2xl text-neutral-900 shadow backdrop-blur-xl transition hover:bg-white/90 hover:shadow-md active:scale-95"
