+++ apps/light-client/app/page.tsx
@@ -7,7 +7,7 @@ import { fetchMetadata } from "./fetchMetadata";
 const SPACE_ID = 13; // Change this to your space ID
 
 export const metadata: Metadata = {
-  title: `Space ${SPACE_ID.toString(16).padStart(4, "0x")} // The Wired`,
+  title: `Space ${SPACE_ID.toString(16).padStart(4, "0x")} / The Wired`,
 };
 
 export default async function Home() {
