+++ apps/light-client/package.json
@@ -15,7 +15,8 @@
     "ethers": "^5.7.2",
     "next": "13.2.4",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-icons": "^4.8.0"
   },
   "devDependencies": {
     "@types/node": "^18.15.3",
