+++ packages/react-client/src/hooks/useEngine.ts
@@ -1,6 +1,6 @@
 import { useContext } from "react";
 
-import { ClientContext } from "../Client";
+import { ClientContext } from "../components/Client";
 
 export function useEngine() {
   return useContext(ClientContext).engine;
