+++ packages/react-client/src/hooks/useLoading.ts
@@ -1,6 +1,6 @@
 import { useContext } from "react";
 
-import { ClientContext } from "../Client";
+import { ClientContext } from "../components/Client";
 
 export function useLoading() {
   const { loadingProgress, loadingText } = useContext(ClientContext);
