+++ packages/react-client/src/hooks/useMic.ts
@@ -1,6 +1,6 @@
 import { useContext } from "react";
 
-import { ClientContext } from "../Client";
+import { ClientContext } from "../components/Client";
 
 export function useMic() {
   const { micEnabled, setMicEnabled, micTrack, setMicTrack } = useContext(ClientContext);
