+++ packages/react-client/src/hooks/useSpaceId.ts
@@ -1,6 +1,6 @@
 import { useContext } from "react";
 
-import { ClientContext } from "../Client";
+import { ClientContext } from "../components/Client";
 
 export function useSpaceId() {
   return useContext(ClientContext).spaceId;
